[Fix](nereids) collect trival-agg instead of all aggs in NormalizeRepeat
